refactor image and text generator
